basic OperationEngine service: Service class for binaries running on hosts
Create the basic Service class for binaries running
on hosts.
A service takes a manager and enables rpc by listening
to queues based on topic. It also periodically runs
tasks on the manager(optional) and reports it state
to the database services table.
Also add some unit tests for Service class.

Partial-Bug: #1527097
Change-Id: I241c4757d0b2d1880d1a1a59cce007ca9d1037c7