Encapsulate smaug client in Operation Engine
According to the design, all the operations which will access
'protection service' module should invoke the API of Smaug instead.
Take the protect operation for example, it will POST the requst
of '/v1/providers/{provider_id}/checkpoints' to execute the
protect operation. Therefore, I encapsulate the client of Smaug
to send the all kinds of requsts conveniently.

Change-Id: Ifc06e49df764d6b17b08f2e9d3f2cb9c64e205c0
Closes-Bug: #1575569