Add trust_id to db table 'scheduled_operation_state'
According to the design, all the operations which will access
'protection service' module should invoke the API of Smaug instead.
Take the protect operation for example, it will POST the requst
of '/v1/providers/{provider_id}/checkpoints' to execute the
protect operation. Therefore, I encapsulate the client of Smaug
to send the all kinds of requsts conveniently.

As we know, operation may happen at any time. we should have a
correct token of Keystone to send quests when the operation is
triggered. When creating a new scheduled operation, I create
a 'trust' in Keystone in the method
smaug_client.py/AuthManager.add_operation. When we need call
the API of Smaug, we create an instance of Smaug client with
a correct token generated by the 'trust'.

Change-Id: If01e481b0c7ec5cb4dd222f1b73502b5504e73d7
Closes-Bug: #1587789