Correct logging translation markers
There were a few log strings that were not using the appropriate
translation marker for their given log level.

Also fixed a couple cases of passing in an exception to a LOG.exception
call. LOG.exception will automatically log whatever exception is in
scope, so switched these to be the equivalent LOG.error levels so the
exceptions do not get logged twice.

Change-Id: If1be4661174dfcf30567b5093bd5373c2c275cf5