Add exception message translations
There were a few cases where strings sent to exceptions were not
being translated. According to the i18n guidelines, any strings
used for exceptions should use the _() marker.

This addresses the obvious cases where a string was being passed
in the exception constructor. There may be other cases where strings
are being assigned to a variable first before being passed in that
still need to be fixed.

Change-Id: I18e0a9bb9bc8ca06d6774ef992e500594df3d4b1