Stabilize Protection Plugin API
Protection plugin now needs to implement methods for each operation
(protect, delete, restore). Each protection plugin is created in the
context of the flow, while each operation is created in the context of
the resource.

A protection plugin is responsible for returning an Operation class for
each operation. Such Operation class defines the behavior of the
protection plugin by implementing these optional hooks:

- on_prepare_begin hook is invoked before any hook of this resource and
  dependent resources has begun
- on_prepare_finish hook is invoked after any prepare hooks of dependent
  resources are complete.
- on_main hook is invoked after the resource prepare hooks are complete
- on_complete hook is invoked once the resource's main hook is
  complete, and the dependent resources' on_complete hooks are complete

HeatTemplate is now created and supplied from a task instead of being
created and passed to resource tasks before the restore operation began.

Change-Id: I847eec6990b2d24a66a12542d242fbfb682272fe
Co-Authored-By: Saggi Mizrahi <saggi.mizrahi@huawei.com>
Implements: blueprint protection-plugin-is-design