Cinder Protection Plugin: Backup from snapshot
Cinder protection plugin used to backup the cinder volume by creating a
backup from the volume, whether it is attached or not.
This patch makes the protection plugin configurable: whether to create
the backup from the volume (using force), or create a snapshot from the
volume, create a backup from the snapshot, and delete the snapshot.

References:
  https://specs.openstack.org/openstack/cinder-specs/specs/mitaka/backup-snapshots.html
  https://blueprints.launchpad.net/cinder/+spec/backup-snapshots
  https://review.openstack.org/#/c/232305/
  https://review.openstack.org/#/c/243406/

Change-Id: Iae600a9537cd2ec7e703f28db82967c47dc3eedb