Support clients with privileged user in karbor
In some cases, Karbor services or protection plugins require
clients(cinder/nova) with admin authority to call the RESTAPI
with the rule for admin.

For example: Nova protection plugins from vendor need Nova
client with admin authority to call the RESTAPI about
hypervisor. The tenant karbor can be used to initialize
the clients with admin authority.

Change-Id: I0c59ebabd226f63369ef1038f6be17684491d632
Closes-Bug: #1650239