Add trigger verification in API
Currently, we use 'cast' instead of 'call' to send request
from api to operationengine when creating a new trigger.
With 'cast', it will not give a return value to api even the
trigger is invalid, so users can create triggers successfully
with using invalid triggers.

This patch add trigger verification in API side. Before send
create request to operationengine, API send a verify request
to operationengine first, which type is 'call', to make sure
the trigger will be created is valid.

Change-Id: Iea734e1b4a63f6e35afe32cbbb2b60f79d6db085
Closes-Bug: #1736657