Fix listing checkpoints by plan error with FS bank
When listing checkpoints by plan with using FS bank,
it will fail with 'could not be found' error. The reason
is, in the bank, checkpoints by plan are stored as
"<bank_path>/karbor/indices/by-plan/<plan_id>/<project_id>
/<date>/<checkpoint>", but the file bank plugin only list
the <date> level, not include the checkpoint level. This
patch will fix it.

Closes-Bug: #1802815

Change-Id: I8c1356269277647490925e0181e2cd1dfca1fa4e