Update using storyboard in README
We have started using storyboard instead of launchpad to
do bugs and features tracking. There are still some bugs
are remained in the launchpad, we will move part of them
to storyboard. The new bugs and features should locate
on storyboard now, so this patch update the README to let
the developers know that.
Change-Id: I4fd908e99157e7f477ee181306c94e90aff586cd